requirements-test.txt
# 테스트 관련 패키지
coverage==7.3.2
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# 코드 품질 및 린팅
flake8==6.1.0
black==23.11.0
isort==5.12.0

# 보안 검사
bandit==1.7.5
safety==2.3.5

# 성능 테스트
locust==2.17.0

# 테스트 데이터베이스
django-test-plus==2.2.2
